#4 fix JMenuItemDriver (and therefore it's tests).
With e9ba8b80f606571a8cc20a0704bc995924818c09 the hierarchy of the
component fixtures and drivers has been changed.

With doing so the method click() from the JMenuItemFixture was removed
and only the method of the ComponentFixture was used. If somebody now
called click() on a JMenuItemFixture this call didn't call
JMenuItemDriver#click(JMenuItem) but instead
JMenuItemDriver#click(Component).

The idea of this fix is to hide the complexity of the driver and simply
override the click(Component) in the driver. Then decide which method to
call from the JMenuItemDriver.