#4 add ToolkitClickRecorder to listen for mouse events early when the toolkit awt event listeners are invoked.
Since some popups are shown and it's tested that the right click has
happened, there is a need for this ToolkitClickRecorder. Because without
it, the mouse events may never been recorded.

Normal ClickRecorder is added as a MouseListener. But MouseListeners
don't run when the event is consumed by the toolkit event listeners. The
popup is shown by a toolkit event listener that consumes the MouseEvent.
Therefore the MouseEvent is never noticed by the ClickRecorder.