#4 fix tests showing a popup by using a ToolkitClickRecorder.
The BasicLookAndFeel registers an AWTEventListener to show the popup
menues. This consumes the MouseEvents that trigger the popup menu. The
MouseEvents therefore never occur in >normal< MouseListeners attached to
the Component the click was made on.

Therefore the ToolkitClickRecorder has been invented. It is invoked even
if the event is consumed.