#2 Early exit #pause(Condition[], long) when one condition is null.
A NullPointerException should be thrown when one Condition of the passed
array is null. Previously if all conditions before a null condition
always returned false, the NullPointerException only occurred after the
timeout of the method.

Now before waiting for the conditions to pass a quick null-check is done
so that it can throw the NullPointerException quickly.