#4 #2 Remove useless test case.
The removed test case tested a behaviour reported in BUG 228
(https://code.google.com/p/fest/issues/detail?id=228). The problem was
that the test blocked when showing a modal dialog.

Reading the JavaDoc of Dialog ("setVisible(true): If the dialog is not
already visible, this call will not return until the dialog is hidden by
calling setVisible(false) or dispose.") this is the expected behaviour.

So this test will always block when calling #setVisible(true) and
therefore is IMO useless.