Let Locale for keystrokes be defined per system property.
Previously the JVM Locale has been used for determining the correct
keystroke mapping provider. This fails if the user selects another
Locale for the keyboard than for the JVM.

Now InputContext#getLocale() is used to determine the keyboard Locale.

Since this method is OS dependend the fallback is still the JVM Locale.
For convenience you can also set the Locale with system property
`assertj.swing.keyboard.locale`.

So the Locale is determined as follows:

1. System Property
2. If not set, InputContext#getLocale()
3. If still not set, JVM Locale