Fixed TextNG before and after methods when running with groups
When running a test group in TestNG, before and after methods are only
executed if the "group" parameter of the annotation makes them part
of the executed group.

The before and after methods in AssertJSwingTestngTestCase always have
to run regardless of the test group. This is achieved by setting the
"alwaysRun" parameter of the before/after annotations to true.