Made AbstractComponentFixture.robot() public
It is desirable to be able to factor out code that operates on
component fixtures without having to pass the Robot as a separate
argument.

Another use case: In Kotlin, you can define extension functions on
component fixtures to add custom functionality. Often a robot is
required. With this change you could define:

    fun AbstractComponentFixture<*, *, *>.clickTopLeftCorner() {
        robot().click(target(), Point(0, 0))
    }